backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multi-agent-backend"
version = "0.1.0"
description = "FastAPI + LangGraph orchestrator for multi-agent hackathon demo"
authors = [{ name = "Hackathon Team" }]
dependencies = [
  "fastapi>=0.112.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "langchain>=0.2.11",
  "langgraph>=0.2.3",
  "langchain-openai>=0.1.20",
  "langchain-community>=0.2.10",
  "langchain-ollama>=0.1.0",
  "langchain-google-genai>=1.0.0",
  "langchain-text-splitters>=1.0.0",
  "chromadb>=0.5.0",
  "python-multipart>=0.0.9",
  "pypdf>=4.0.0",
  "python-pptx>=0.6.21",
  "openpyxl>=3.1.0",
  "python-docx>=1.1.2",
  "tiktoken>=0.7.0",
  "httpx>=0.27.0",
  "redis>=5.0.0",
  "sqlalchemy>=2.0.32"
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "httpx",
  "ruff"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uvicorn]
host = "0.0.0.0"
port = 8000

[tool.ruff]
line-length = 100
target-version = "py310"

